crypto: msm: remove wakeup lock in qcrypto driver
When log server still runs and try to enable CE clock just before sleep,
wakeup lock will prevent the device sleep. And the bam/crypto is a not a
wakeupable device. As we have a state machine in crypto driver, and it
is able to control suspend behavior well, we do not need to use wakeup
lock in qcrypto driver.

Change-Id: I660818ec6fad4051db08c10a76d5d1b408bece7e
Signed-off-by: Zhen Kong <zkong@codeaurora.org>
Signed-off-by: CheckYourScreen <nimitmehta95@gmail.com>