vfs: Work around NULL pointer dereference in d_path()
All of a sudden, path->dentry is frequently being NULL when the system is
going down, causing a kernel panic when the user attempts to power down
or reboot the device.

Add a check on path->dentry in order to avoid the NULL pointer dereference.
This behavior is very suspicious, especially since this bug has never
occurred before, so this should be revisited later.

Signed-off-by: Sultanxda <sultanxda@gmail.com>
Signed-off-by: CheckYourScreen <nimitmehta95@gmail.com>