devfreq: Add a Governor white-list
Many of the devfreq governors compiled in are unstable/incomplete, so we
don't want users switching to them. Compiling out the unused governors is
messy because many other drivers might link to them, and there are not proper

Add a white list of the only governors intended for use to the governor
sysfs node in order to prevent users from enabling governors that are unstable and causes reboot.

Signed-off-by: CheckYourScreen <nimitmehta95@gmail.com>